sched/schedutil: Introduce a power aware capacity limiter
We removed the absurd hotplugging mechanism on the Exynos 9810,
however we can't leave the kernel as it is if we want to use the
higher frequencies reasonably.

The hotplugging mechanism worked by reactively monitoring to the
runqueues in cfs and then adjusting the online cores and frequencies.
This murdered performance as any execution noise would force the
cores onto the lower frequencies, similarly, having only one core
online also murdered performance when there was a burst of threads.

Instead of reactively controlling frequency I propose to preemptively
calculate if we're allowed to go to a higher frequency based on the
average load of the CPUs and a defined target power limit.

The mechanism takes the busy state capacity and power tables from
the current CPU scheduler domain which are also used for energy aware
scheduling. Before calculating next_freq in schedutil, we parse the
cost tables and calculate the power for the domain/cpus. If the total
power exceeds the defined limit, we simply cap the capacity to the
next capacity level/frequency which comes under the power_limit.
This mechanism thus still allows single-threaded workloads to reach
peak frequencies, while proactively limiting frequencies in
multi-threaded workloads. We avoid forced migrations and the performance
deficit in interactive ST workloads is nearly completely eliminated
when compared to the hotplugging driver.

Signed-off-by: FAROVITUS <farovitus@gmail.com>

 Conflicts:
	kernel/sched/cpufreq_schedutil.c